requirements.txt
anyio==4.8.0
asgiref==3.8.1
certifi==2025.1.31
Django==5.1.6
django-jazzmin==3.0.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
pillow==11.1.0
psycopg2-binary==2.9.10
python-decouple==3.8
python-telegram-bot==21.10
sniffio==1.3.1
sqlparse==0.5.3
